requirements.txt
streamlit==1.25.0
pandas==2.0.3
numpy==1.24.3
plotly==5.15.0
openpyxl==3.1.2
pillow==9.5.0